+++ R/effective_sample.R
@@ -5,8 +5,8 @@
 #' samples. A higher ESS indicates more reliable estimates. For most
 #' applications, an effective sample size greater than 1,000 is sufficient for
 #' stable estimates (Bürkner, 2017). This function returns the effective sample
-#' size (ESS) for various Bayesian model objects. For `brmsfit` objects, the
-#' returned ESS corresponds to the bulk-ESS (and the tail-ESS is also returned).
+#' size (ESS) for various Bayesian model objects. For `brmsfit`, `stanreg`, and
+#' `stanfit` objects, both **bulk-ESS** and **tail-ESS** are returned.
 #'
 #' @param model A `stanreg`, `stanfit`, `brmsfit`, `blavaan`, or `MCMCglmm` object.
 #' @param ... Currently not used.
@@ -14,7 +14,8 @@
 #'
 #' @inheritSection hdi Model components
 #'
-#' @return A data frame with two columns: Parameter name and effective sample size (ESS).
+#' @return A data frame with columns for the parameter name, bulk ESS (`ESS`),
+#'   and (where available) tail ESS (`ESS_tail`).
 #'
 #' @details
 #'   - **Effective Sample (ESS)** should be as large as possible, altough
@@ -84,11 +85,13 @@ effective_sample.default <- function(model, ...) {
 
 #' @rdname effective_sample
 #' @export
-effective_sample.brmsfit <- function(model,
-                                     effects = "fixed",
-                                     component = "conditional",
-                                     parameters = NULL,
-                                     ...) {
+effective_sample.brmsfit <- function(
+  model,
+  effects = "fixed",
+  component = "conditional",
+  parameters = NULL,
+  ...
+) {
   pars <- insight::find_parameters(
     model,
     effects = effects,
@@ -105,7 +108,7 @@ effective_sample.brmsfit <- function(model,
 
   data.frame(
     Parameter = idx$variable[rows_to_keep],
-    ESS = round(idx[rows_to_keep, "ess_bulk"]),
+    ESS_bulk = round(idx[rows_to_keep, "ess_bulk"]),
     ESS_tail = round(idx[rows_to_keep, "ess_tail"]),
     stringsAsFactors = FALSE,
     row.names = NULL
@@ -114,11 +117,13 @@ effective_sample.brmsfit <- function(model,
 
 
 #' @export
-effective_sample.stanreg <- function(model,
-                                     effects = "fixed",
-                                     component = "location",
-                                     parameters = NULL,
-                                     ...) {
+effective_sample.stanreg <- function(
+  model,
+  effects = "fixed",
+  component = "location",
+  parameters = NULL,
+  ...
+) {
   effective_sample.brmsfit(
     model,
     effects = effects,
@@ -130,11 +135,13 @@ effective_sample.stanreg <- function(model,
 
 
 #' @export
-effective_sample.stanmvreg <- function(model,
-                                       effects = "fixed",
-                                       component = "location",
-                                       parameters = NULL,
-                                       ...) {
+effective_sample.stanmvreg <- function(
+  model,
+  effects = "fixed",
+  component = "location",
+  parameters = NULL,
+  ...
+) {
   pars <- insight::get_parameters(
     model,
     effects = effects,
@@ -150,7 +157,7 @@ effective_sample.stanmvreg <- function(model,
 
   data.frame(
     Parameter = idx$variable[rows_to_keep],
-    ESS = round(idx[rows_to_keep, "ess_bulk"]),
+    ESS_bulk = round(idx[rows_to_keep, "ess_bulk"]),
     ESS_tail = round(idx[rows_to_keep, "ess_tail"]),
     stringsAsFactors = FALSE,
     row.names = NULL
@@ -159,24 +166,23 @@ effective_sample.stanmvreg <- function(model,
 
 
 #' @export
-effective_sample.stanfit <- function(model,
-                                     effects = "fixed",
-                                     parameters = NULL,
-                                     ...) {
+effective_sample.stanfit <- function(model, effects = "fixed", parameters = NULL, ...) {
   pars <- insight::get_parameters(
     model,
     effects = effects,
     parameters = parameters
   )
 
-  insight::check_if_installed("rstan")
-
-  s <- as.data.frame(rstan::summary(model)$summary)
-  s <- s[rownames(s) %in% colnames(pars), ]
+  insight::check_if_installed("posterior")
+  idx <- as.data.frame(posterior::summarise_draws(model))
+  rows_to_keep <- idx$variable %in% colnames(pars)
+  # ess_*() functions are defined in:
+  # https://github.com/stan-dev/posterior/blob/master/R/convergence.R
 
   data.frame(
-    Parameter = rownames(s),
-    ESS = s[["n_eff"]],
+    Parameter = idx$variable[rows_to_keep],
+    ESS_bulk = round(idx[rows_to_keep, "ess_bulk"]),
+    ESS_tail = round(idx[rows_to_keep, "ess_tail"]),
     stringsAsFactors = FALSE,
     row.names = NULL
   )
@@ -203,10 +209,7 @@ effective_sample.blavaan <- function(model, parameters = NULL, ...) {
 
 
 #' @export
-effective_sample.MCMCglmm <- function(model,
-                                      effects = "fixed",
-                                      parameters = NULL,
-                                      ...) {
+effective_sample.MCMCglmm <- function(model, effects = "fixed", parameters = NULL, ...) {
   pars <- insight::get_parameters(
     model,
     effects = effects,
@@ -225,12 +228,15 @@ effective_sample.MCMCglmm <- function(model,
   )
 
   if (nrow(s.random) > 0L) {
-    es <- rbind(es, data.frame(
-      Parameter = rownames(s.random),
-      ESS = round(s.random[["eff.samp"]]),
-      stringsAsFactors = FALSE,
-      row.names = NULL
-    ))
+    es <- rbind(
+      es,
+      data.frame(
+        Parameter = rownames(s.random),
+        ESS = round(s.random[["eff.samp"]]),
+        stringsAsFactors = FALSE,
+        row.names = NULL
+      )
+    )
   }
 
   es[match(pars[[1]], es$Parameter), ]